Correct reversal of regexps for qNames.
The previous implementation appears to be broken, as it (for example) doesn't allow for numbers to appear in the name part of a qName (e.g. bflc_name00MatchKey). See https://www.w3.org/TR/REC-xml/#NT-Name for the formal definition.

This change also required an update to the encoder tests.